+++ src/lib/components/NotebookDetail.svelte
@@ -1156,7 +1156,6 @@
 							name="content"
 							rows={6}
 							required
-							start="both"
 							todos={todoRefs}
 							placeholder={t('notebookDetail.writeANoteAbout', { title: notebook.title })}
 						/>
@@ -1659,7 +1658,6 @@
 								name="content"
 								rows={8}
 								required
-								start="both"
 								todos={todoRefs}
 							/>
 							<PictureAttach target={editBox} />